+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Remove deprecated PodSecurityPolicy.
+
 ### Changed
 
 - Update Arch PKGBUILD for 0.9.0.